gl_func_0006FB54: unsigned delta cast in 64-bit add, 36.8%->42.6%
The 64-bit accumulate `base + delta` sign-extended the signed `delta` into the high
word (extra sra). Casting `(unsigned int)delta` in the add makes it zero-extend (adds
to low + carry only), matching the target. +5.8pp. Residual is the 64-bit-add spill/
register structure (3 globals all D_00000000 @ 0/0/4, no-CSE re-materialized luis).

+++ src/game_libs/game_libs_post.c
@@ -40242,7 +40242,7 @@ long long gl_func_0006FB54(int a0) {
     long long base = ((long long)*(int *)&D_00000000 << 32) |
                      (unsigned int)*(int *)((char *)&D_00000000 + 4);
     gl_func_00000000(r0);
-    return base + delta;
+    return base + (unsigned int)delta;
 }
 #else
 INCLUDE_ASM("asm/nonmatchings/game_libs/game_libs", gl_func_0006FB54);
